Rename internal `Request` to `ClientRequestStream`
